+++ edzed/blocklib/fsms.py
@@ -9,6 +9,7 @@ Home: https://github.com/xitop/edzed/
 from __future__ import annotations
 
 from .. import fsm
+from .. import utils
 
 __all__ = ['Timer']
 
@@ -31,6 +32,11 @@ class Timer(fsm.FSM):
         )
 
     def __init__(self, *args, restartable: bool = True, **kwargs):
+        if 't_period' in kwargs:
+            if ('t_on' in kwargs or 't_off' in kwargs):
+                raise TypeError("t_period and t_on/t_off are mutually exclusive.")
+            period = utils.time_period(kwargs.pop('t_period'))
+            kwargs['t_on'] = kwargs['t_off'] = period / 2
         super().__init__(*args, **kwargs)
         self._restartable = bool(restartable)